Fix 32-bit test failure: assert Int, not Int64
The "Date and time columns" testitem checked that a non-temporal column
survives untouched with `typeof(rows[1].n) == Int64`, but the column is
written as [1, 2], whose eltype is Int -- and Int is Int32 on the x86 CI
legs. All five failures in the last run were this one assertion, on
ubuntu-x86 and windows-x86; every x64 leg passed.

Verified on 32-bit Julia locally (julia +release~x86, Sys.WORD_SIZE=32,
Int=Int32) rather than only on x64: 45/45 tests pass there, and the x64
run still passes.

+++ test/test_featherfiles.jl
@@ -127,7 +127,9 @@ end
     @test rows[1].dt == DateTime(2020, 1, 2, 3, 4, 5)
     @test typeof(rows[1].t) == Time
     @test rows[1].t == Time(1, 2, 3)
-    @test typeof(rows[1].n) == Int64      # non-temporal columns untouched
+    # Int, not Int64: the column was written as [1, 2], whose eltype is Int, and that is
+    # Int32 on the 32-bit CI legs.
+    @test typeof(rows[1].n) == Int        # non-temporal columns untouched
 
     cols = TableTraits.get_columns_copy_using_missing(load(filename))
     @test eltype(cols.d) == Date
